Refactored style functions to own file. Better support for line styles.
