+++ bundles/mapping/mapmodule/oskariStyle/constants.js
@@ -1,0 +1,36 @@
+// OSKARI STYLE CONSTANTS
+// These are mentioned in the api documentation
+
+export const LINE_DASH = {
+    DASH: 'dash',
+    DASHDOT: 'dashdot',
+    DOT: 'dot',
+    LONGDASH: 'longdash',
+    LONGDASHDOT: 'longdashdot',
+    SOLID: 'solid'
+};
+
+export const LINE_JOIN = {
+    MITRE: 'mitre',
+    ROUND: 'round',
+    BEVEL: 'bevel'
+};
+
+export const LINE_CAP = {
+    BUTT: 'butt',
+    ROUND: 'round'
+};
+
+export const EFFECT = {
+    DARKEN: 'darken',
+    LIGHTEN: 'lighten'
+};
+
+export const FILL_STYLE = {
+    THIN_DIAGONAL: 0,
+    THICK_DIAGONAL: 1,
+    THIN_HORIZONTAL: 2,
+    THICK_HORIZONTAL: 3,
+    TRANSPAREN: 4,
+    SOLID: 5
+};
